+++ src/REvoDesign/bootstrap/__init__.py
@@ -5,6 +5,7 @@ import os
 
 import hydra
 
+from .. import Qt
 from .set_config import (experiment_config, reload_config_file,
                          save_configuration, set_cache_dir,
                          set_REvoDesign_config_file)
@@ -21,6 +22,7 @@ hydra.initialize_config_dir(
 EXPERIMENTS_CONFIG_DIR = experiment_config()
 
 __all__ = [
+    'Qt',
     "experiment_config",
     "set_REvoDesign_config_file",
     "reload_config_file",