+++ src/REvoDesign/shortcuts/shortcuts.py
@@ -16,6 +16,9 @@ from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
 from immutabledict import immutabledict
 from pymol import cmd, util
+from RosettaPy import (Rosetta, RosettaEnergyUnitAnalyser,
+                       RosettaScriptsVariableGroup)
+from RosettaPy.app.fastrelax import FastRelax, get_relax_scripts_from_db
 from RosettaPy.app.pross import PROSS
 from RosettaPy.app.rosettaligand import RosettaLigand
 from RosettaPy.app.utils.smiles2param import SmallMoleculeParamsGenerator
@@ -795,3 +798,99 @@ def shortcut_pross(
     best_pdb_path = pross.design(filters=filters, refined_pdb=best_refined, filterscan_dir=filterscan_dir)
 
     logging.info(f"PROSS design finished. Best pdb: {best_pdb_path}")
+
+
+def shortcut_fast_relax(
+        pdb: str,
+        relax_script: str,
+        nstruct: int = 4,
+        dualspace: bool = False,
+        default_repeats: int = 3,
+        job_id: str = 'fastrelax',
+        save_dir: str = 'relaxed',
+        relax_opts: Optional[List[Union[str, RosettaScriptsVariableGroup]]] = None,
+
+
+):
+    bus = ConfigBus()
+
+    node_config = read_rosetta_node_config()
+
+    class FastRelaxOpts(FastRelax):
+        def run(self, nstruct: int = 8, default_repeats: int = 15,
+                relax_opts: Optional[List[Union[str, RosettaScriptsVariableGroup]]] = None) -> RosettaEnergyUnitAnalyser:
+            """
+            Runs the fast relaxation process using the specified parameters.
+
+            Args:
+                nstruct (int, optional): The number of structures to generate. Defaults to 8.
+                default_repeats (int, optional): The default number of repeats for relaxation. Defaults to 15.
+
+            Returns:
+                RosettaEnergyUnitAnalyser: An object for analyzing the energy units of the generated structures.
+            """
+            if relax_opts is None:
+                relax_opts = []
+            # Configure and run Rosetta for fast relaxation
+            rosetta = Rosetta(
+                bin="relax",
+                opts=[
+                    "-in:file:s",
+                    os.path.abspath(self.pdb),
+                    "-relax:script",
+                    self.relax_script,
+                    "-relax:default_repeats",
+                    str(default_repeats),
+                    "-out:prefix",
+                    f"{self.instance}_fastrelax_",
+                    "-out:file:scorefile",
+                    f"{self.instance}_fastrelax.sc",
+                    "-score:weights",
+                    "ref2015_cart" if self.dualspace else "ref2015",
+                    "-relax:dualspace",
+                    "true" if self.dualspace else "false",
+                ] + relax_opts,
+                save_all_together=True,
+                output_dir=os.path.join(self.save_dir, self.job_id),
+                job_id=f"fastrelax_{self.instance}_{os.path.basename(self.relax_script)}",
+                run_node=self.node,
+                verbose=True
+            )
+
+            with timing("FastRelax"):
+                rosetta.run(nstruct=nstruct)
+
+            analyser = RosettaEnergyUnitAnalyser(rosetta.output_scorefile_dir)
+            best_hit = analyser.best_decoy
+            pdb_path = os.path.join(rosetta.output_pdb_dir, f'{best_hit["decoy"]}.pdb')
+
+            print("Analysis of the best decoy:")
+            print("-" * 79)
+            print(analyser.df.sort_values(by=analyser.score_term))
+
+            print("-" * 79)
+
+            print(f'Best Hit on this FastRelax run: {best_hit["decoy"]} - {best_hit["score"]}: {pdb_path}')
+            return analyser
+
+    fast_relax = FastRelaxOpts(
+        pdb=pdb,
+        relax_script=relax_script,
+        dualspace=dualspace,
+        job_id=job_id,
+        save_dir=save_dir,
+        node=node_picker(
+            node_type=bus.get_value('rosetta.node_hint', str, reject_none=True),  # type: ignore
+            nproc=bus.get_value('ui.header_panel.nproc', int, reject_none=True),
+            **node_config)
+    )
+
+    analyser = fast_relax.run(
+        nstruct=nstruct,
+        default_repeats=default_repeats,
+        relax_opts=relax_opts,
+    )
+
+    best_relaxed_decoy = analyser.best_decoy
+
+    logging.info(f"FastRelax finished. Best decoy: {best_relaxed_decoy}")