+++ tests/menu/shortcuts/test_rosetta_tasks.py
@@ -3,7 +3,8 @@ import os
 import pytest
 from RosettaPy.analyser import RosettaEnergyUnitAnalyser
 
-from REvoDesign.shortcuts.shortcuts import (shortcut_pross,
+from REvoDesign.shortcuts.shortcuts import (shortcut_fast_relax,
+                                            shortcut_pross,
                                             shortcut_rosettaligand)
 from tests.conftest import TestWorker
 
@@ -19,9 +20,8 @@ from tests.conftest import TestWorker
     ],
 )
 def test_rosetta_ligand(job_id, start_from, cst, test_worker: TestWorker, test_node_hint):
-    if test_node_hint == 'docker':
-        test_worker.plugin.bus.set_value('rosetta.node_hint', test_node_hint)
-        test_worker.plugin.bus.set_value('rosetta.node_config', {"mpi_available": True, 'prohibit_mpi': False})
+
+    test_worker.inject_rosetta_node_config(test_node_hint)
 
     save_dir = 'rosetta_tests/outputs'
 
@@ -50,9 +50,7 @@ def test_rosetta_ligand(job_id, start_from, cst, test_worker: TestWorker, test_n
 
 # time expensive test
 # def test_pross(test_worker: TestWorker, test_node_hint):
-#     if test_node_hint == 'docker':
-#         test_worker.plugin.bus.set_value('rosetta.node_hint', test_node_hint)
-#         test_worker.plugin.bus.set_value('rosetta.node_config', {"mpi_available": True, 'prohibit_mpi': False})
+#     test_worker.inject_rosetta_node_config(test_node_hint)
 
 #     save_dir = 'rosetta_tests/outputs'
 #     shortcut_pross(
@@ -97,3 +95,47 @@ def test_rosetta_ligand(job_id, start_from, cst, test_worker: TestWorker, test_n
 #     analyser = RosettaEnergyUnitAnalyser(os.path.join(design_dir, 'scorefile'))
 #     assert not analyser.df.empty, 'Scorefile should be loaded and analysed'
 #     assert analyser.df.shape[0] == 8, 'Scorefile should contain exact 8 rows'
+
+
+@pytest.mark.serial
+# a variant test from RosettaPy's short app tests
+@pytest.mark.parametrize(
+    "job_id,pdb,dualspace,ligand",
+    [
+        ['mono', '../tests/data/3fap_hf3_A_short.pdb', False, ''],
+        ['mono_dualspace', '../tests/data/3fap_hf3_A_short.pdb', True, ''],
+        ['w_ligand', '../tests/data/pdb/3fap_hf3_A_short_lig.pdb', False, '../tests/data/lig/lig.fa.params'],
+    ],
+)
+def test_fast_relax(job_id, pdb, dualspace, ligand, test_worker: TestWorker, test_node_hint):
+
+    test_worker.inject_rosetta_node_config(test_node_hint)
+
+    save_dir = 'rosetta_tests/outputs/fastrelax'
+    relax_script = 'MonomerRelax2019'
+
+    relax_opts = []
+    if ligand:
+        relax_opts.extend(['--extra_res_fa', os.path.abspath(ligand)])
+
+    shortcut_fast_relax(
+        pdb=os.path.abspath(pdb),
+        relax_script=relax_script,
+        dualspace=dualspace,
+        job_id=job_id,
+        save_dir=save_dir,
+        nstruct=4,
+        default_repeats=3,
+        relax_opts=relax_opts,
+    )
+    pdb_bn = os.path.basename(pdb)[:-4]
+
+    run_dir = os.path.join(save_dir, f'{job_id}/fastrelax_{pdb_bn}_{relax_script}')
+    assert os.path.isdir(os.path.join(run_dir, 'all')), f"{run_dir}/all does not exist"
+    assert [x for x in os.listdir(os.path.join(run_dir, 'all')) if x.endswith('.pdb')
+            ], f'{run_dir}/all should contain pdb files'
+    assert [x for x in os.listdir(os.path.join(run_dir, 'all')) if x.endswith(
+        '.sc')], f'{run_dir}/all should contain score files ends with .sc'
+
+    analyser = RosettaEnergyUnitAnalyser(os.path.join(run_dir, 'all'))
+    assert not analyser.df.empty, 'Scorefile should be loaded and analysed'